[Breaking Change] B3Propagator.Builder -> B3PropagatorBuilder
